fixed RoiTools.computeTiledROIs
Noticed that by column option did not work (the bottom tiles would never be generated) and that the geometryLocal was not different from geometry (because of typo in the index for row/columnParents).

Made some changes, hope it works now.

Revert "fixed RoiTools.computeTiledROIs"

This reverts commit b5218a640600704f1278dfee1ce275549f22424b.

ignore

fixed RoiTools.computeTiledROIs for very large annotations (full image)

Noticed that byColumn option did not work (the bottom tiles would never be generated when only using prepared geometries) and that the geometryLocal was not different from geometry (because of typo in the index for row/columnParents).

Made some changes, hope they work.

Revert "ignore"

This reverts commit 4bab614a20d62b66e4a15eb1b737e96d72a2358a.