Add show javadocs command to script editor
Found under 'Help > Show Javadocs'.
This searches for a local jar file containing the javadocs.
It also attempts to find all relevant javadoc jars when running QuPath using `gradlew run` or from an IDE.

If needed, a system property can be used to set the javadoc path or URL -Djavadoc=/path/to/javadocs