Add support for custom styles
This adds support to add custom styles by adding .css files to the a css subdirectory inside the user directory (which is where extensions go, usually at ~/QuPath/v0.4/ or similar).

If the custom css directory exists, a WatchService is created to check for changes. This means that users can add/remove/modify .css files in the directory, and see live updates (albeit with some seconds of delay, at least on macOS).

To make this work, the PreferencePane needed to be updated to directly use any ObservableList with combo boxes.
A new `PathPrefs.getCssStylesPath()` was also added.

Additionally, IconFactory.DuplicableGlyph was improved to allow glyphs to respond to the `-fx-text-fill` property, and this in turn means that the toolbar icons can now be styled by css (rather than being fixed to a neutral gray).

Finally, main.css was introduced to enable more QuPath-specific styling, even when using the default modena.css with JavaFX. This is particularly relevant for defining several colors reused by the script editor.