Add more map-like syntax to MeasurementList
This mostly provides shorter alternatives to `putMeasurement(String, double)` and `getMeasurementValue(String)` (i.e. `put(String)` and `get(String)` respectively).

The intention is that it is easier to switch between using a MeasurementList or a map since the key methods have the same syntax - although only a small subset of Map methods are implemented.

The reasons to use the list are
- required for backwards-compatible serialization support
- works with primitive values

The reasons to use the map are:
- familiar API
- support for null to be returned for missing values (rather than NaN)
- more convenient with scripting
- no need to call .close() for efficiency (this is carried out internally)