Support background estimate without reconstruction
Make opening by reconstruction optional as a background subtraction method.

This is intended help address issues like
* https://github.com/petebankhead/qupath/issues/22
* https://github.com/qupath/qupath/issues/80

Keeping the option on (the default) should give the same results as previous QuPath versions. Turning it on should improve consistency when detection is applied to the same pixels within different ROIs (i.e. when they are tiled differently for detection) because all the background calculations are then local.