Support '-Dnoextensions' system property
This makes it possible to start QuPath without loading any (UI-based) extensions. This doesn't affect loading ImageServerBuilder implementations on the classpath.

Also avoid relying on `new QP()` within the processing extension, which was causing Gson type adapters not to be loaded if the extension was turned off.

Need to confirm this doesn't break scripts that rely on extensions...