Better exception handling deserializing qpdata
Addresses
* https://github.com/qupath/qupath/issues/1252

To test this, we need a project that contains a broken (truncated) .qpdata file, and a corresponding complete .qpdata.bkp file.

If we have the last of these, we can create a truncated file in a Groovy script, e.g. with

```groovy
def full = "/path/to/data.qpdata.bkp"
def partial = "/path/to/data.qpdata"
new File(partial).bytes = new File(full).bytes[0..10000]
```

Attempting to serialize this should throw an exception. Previously, this wasn't an `IOException`, and so QuPath didn't automatically switch to the backup.... now it should.