Rename PathTools to GridPaneUtils, dependency bump
This fits better with the `qupath-fx` naming and also the purpose of the class. The non-GridPane-related methods are moved to `FXUtils`.

Meanwhile a `module-info.java` is added to `qupath-fx`.

`checkstyle` gradle plugin (temporarily?) removed from `qupath.common-conventions.gradle` as it wasn't behaving nicely with the new `qupath-fx` module.

There are also lots of new warnings from the `mergedJavadocs` task (which is broken...).

Bio-Formats and DJL dependency updates.