Simplify PathPlugin and PluginRunner
Quite a major refactoring & simplification of how `PathPlugin` implementations are handled.

The original problem was that `PluginRunnerFX` proved to be buggy with detached windows, because the `ImageData` being handled was accessed via `PluginRunner.getImageData()`. In most cases that's fine, but `PluginRunnerFX` was always returning the *current* `ImageData` from QuPath, and if this *changed* before the plugin was complete (and had fired any hierarchy updates /  logged any workflow steps) then exceptions or unexpected behavior could occur.

It might conceivably have happened previously when using multiple viewers, but I don't know of any cases where it did - because QuPath locks the UI when plugins run, and this seemed to be working ok.

The new design greatly simplifies `PluginRunner` to simple run tasks in parallel with progress notification - with some special logic for `PathTask` implementations.

Alongside that,
```java
PathPlugin.runPlugin(PluginRunner pluginRunner, String arg)
```
has been updated to
```java
PathPlugin.runPlugin(PluginRunner pluginRunner, ImageData<T> imageData, String arg)`
```
so that there's no need to request the `ImageData` from the plugin runner - and it's possible to ensure this remains unchanged, no matter what might go on in the UI.

Early indications are that the code is just simpler and more robust. However, it's still a core change that *could* potentially have broken lots of things, so most analysis-related commands ought to be checked (e.g. cell detection, or anything else involving `PathPlugin`). In particular, the ImageJ macro runner might be affected.