Fix ImageJ menus on macOS
Should be a better fix for https://github.com/qupath/qupath/issues/6
Basically, QuPath's use of a system menubar is temporarily blocked while ImageJ is active.
Additionally, when no AWT window is in focus, if the previous window has a menubar it is temporarily blocked until an AWT window comes back into focus. The is needed to prevent key combinations from triggering events both in QuPath and ImageJ.