Simplify (slightly) the drawing tool handlers
The aim is to avoid a memory leak where storing a 'parent object' when drawing (in order to constrain some tools, such as the brush and wand) would inadvertently result in a whole object hierarchy remaining in memory.

This could cause hundreds of MB to retained unnecessarily after an image was closed.

In practice, the issue was likely to be resolved quickly through interacting with the viewer some more, but it would be better to avoid the problem entirely.