Improve ImageDisplay initialization
Throw an `IOException` if `ImageDisplay.setImageData()` fails, and remove the constructor that takes an `ImageData` parameter in favor of adding a static `create()` method instead.

The purpose is to 'fail faster and better' if pixels cannot be read from an image, to avoid putting the viewer into an invalid state.

Previously, failure to open an image (e.g. because JPEGXR decompression was unavailable on Apply Silicon) would lead to an extremely uninformative NPE, and then possibly a lot more exceptions when attempting to interact with the viewer in any way.

When making these changes, the constructors for `QuPathViewer` were changed to no longer take an `ImageData` as a parameter. This was always null anyway (viewers were always created without an image), so makes no real difference to the current code - but the change avoids the possibility of the constructor having to throw an exception if the image couldn't be set. Now the image must be set as a second step, after the viewer has been created.