Refactor OpenSlide JNA, add extension
Hoping this doesn't just make a mess, sorry for the main commits...

Refactored the JNA code to make multiple attempts at loading possible, and also provide a search path. This can now be added in the preferences.

It still needs some improvement (e.g. it very eagerly attempts to load while the user types a path...), but if I set the path to /opt/homebrew/lib then I can use my local Homebrew installation of OpenSlide.

My hope is that it helps with testing potential builds, as well as providing the backup in case a truly portable build remains elusive.