Temporarily reinstate old methods/classes
Bring back several key methods and classes that are used in extensions, and mark them as deprecated while logging a warning the first time they are used.
The purpose is to not break more extensions than necessary, and give a clue as to how to update the code for future compatibility.