Make zoom-to-fit a regular (non-selectable) action
Avoid setting a 'zoom-to-fit' status that affects all viewers, but instead just adjust the zoom/positioning of the current viewer.
This is to avoid the problem where a user might accidentally select the option, and then find themselves unable to zoom/pan in the image until the toggle has been deselected.