Improve use of json with image servers
Make JSONImageServerBuilder genuinely useful by enabling it to read the server.json files stored within projects.

This involves estimating the "builderType" since that wasn't actually serialized within the server.json before.

To simplify things in the future, the server.json now serializes the `ImageServer` and not the `ServerBuilder` - so that more information is present.