Simplify DnnModel implementation
Remove generic parameter and move prediction & blob functions to `AbstractDnnModel`.
The motivations are:
- make it easier to implement `DnnModel`
- make it easier to handle memory management with the DeepJavaLibrary implementation

The second is because memory management seems to work best when using a `Translator` and not working directly with `NDList` inputs and outputs for prediction. This means that it's better to predict 'all at once' rather than across multiple methods (create blob, predict, convert output).