Fix ImageDisplay json channel update bug
Fixes https://github.com/qupath/qupath/issues/1468

There may still be an underlying issue whenever the display is initialized with a new server.

What happened was that `selectedChannels` contained one extra channel that wasn't in `availableChannels`, so this was never removed within the `updateFromJSON` method. The bonus channel had the same name/properties as the in `availableChannels`, but the object itself was different.

The situation 'corrected itself' whenever the image's channels were manipulated directly, not via 'Apply to similar image' changes in a different viewer.

Nevertheless, anything in the selected channels list *should* always be in the available channels list - if that was truly the case, the previous implementation would have worked.