Improve handling of extension installation
Add a choice dialog when installing extensions
through the GUI (ie, by inputting an owner and repo).
First, if
multiple releases are present in the github repo,
ask the user to choose one. Then, if there is more
than one valid (ie, a java archive that's not
sources or javadoc), then ask them to choose one
again. Finally, install as before.