Send cells to ImageJ macro runner
A step towards fixing https://github.com/qupath/qupath/issues/1522

Incomplete, because it behaves inconsistently depending upon whether the macro is being run on a selected object or across all annotations / TMA cores. In the first case, the `ImageDisplay` and `OverlayOptions` may be used, but in the second case they are unavailable.