Avoid converting GeometryCollection to Area
This was proving a major bottleneck with very complex ROIs (e.g. generated by merging all cells in an image), rendering QuPath unusable until there was an out-of-memory error.
The use of JTS' `GeometryCollectionShape` leads to much better performance.
We need to also avoid conversion to Area when creating icons.