Improve ImageServer serialization in projects
Serialize the `ServerBuilder` and not the `ImageServer`, which means we can serialize without needing to build the server itself.

Also support creating an `ImageServer` from a `ServerBuilder` JSON representation directly.
This means that an`ImageServer` can be created with `fromJson(json, ImageServer.class)` applied to the output of both `toJson(obj, ServerBuilder.class)` and `toJson(obj, ImageServer.class)`