Simplify PathIO
Fix the infinite recursion caused by a recent change, and occurring when attempting to open a .qpdata file outside a project.

Also clean up some other parts of the class, and add simpler `readImageData` methods that don't require so many parameters.

Deprecate the previous `readImageData` methods and log warnings when they are called.