Speed up object masking
Addresses https://github.com/qupath/qupath-extension-instanseg/issues/88

It feels awkward to use a static `WeakHashMap` but it seems the easiest way to solve the performance issue without complicating the API with more methods to mask objects (taking `Geometry` or `PreparedGeometry`).

With this change, the original problematic image becomes bound by the inference speed rather than the geometry operations - as it should be.