Hide script editor on close requests
Potential fix for https://github.com/qupath/qupath/issues/1660

This hides the script editor when either pressing the corner button to close the window, or calling 'File > Close editor' - but does *not* close the scripts.

However, when QuPath is being quit entirely, if there are unsaved scripts then the script editor should appear again - with prompts to save.