Format blockquotes in webviews
We can only apply one CSS stylesheet to WebViews - which is used by the `WebViews` class to bind to dark mode changes.

Therefore we need to add any interesting custom formatting to those shared stylesheets.

This PR then adds some provisional support for nicer blockquote formatting, so that this is available elsewhere if needed (note that it's subject to change when someone who is better at css looks at it).