Prompt when quitting ImageJ
Give opportunity to save changes - but in one step, instead of looping through all images (which is the default behavior).

This tries to strike a balance between making it quick and easy to close ImageJ after use, while not making it *so* quick and easy that unsaved changes are lost often.