Better building, javadocs
Support specifying toolchain with Gradle (first) or system property (second) - rather than using a gradle properly only.

This is needed because `includeBuild` requires the extension to respect the toolchain, and it is easier to as a system property that is used everywhere, i.e. within `gradle.properties` add a line such as
```
systemProp.toolchain=23
```
and then this should propagate to any included builds that use the QuPath Gradle plugin v0.2.1 (or later, presumably).

Also allow `include-extra` to have any file extension or none (shortest name is preferred). This is to make it easier to handle in an IDE, e.g. using `include-extra.properties` means that shortcuts to comment/uncomment lines can be used - but we don't want to enforce the use of `.properties` as the extension since it isn't really a properties file (it doesn't contain key-value pairs).

Also fix some javadocs (errors emerged when trying a different toolchain).