'Deselection mode' when pressing Alt
This PR addresses https://github.com/qupath/qupath/issues/1704 by changing 'selection mode' to become 'deselection mode' when the 'Alt' key is pressed.

This is comparable to how drawing tools become 'subtraction' tools when the 'Alt' key is pressed.

Therefore there are 3 different behaviors possible when drawing a ROI with selection mode turned on:
- No keys pressed - reset any current selection, and select any objects within the ROI
- Shift key pressed - add any objects within the ROI to the current selection
- Alt key pressed - remove any objects within the ROI from the current selection

(Currently there is no visual difference in how the ROI is rendered to reflect these different behaviors, because it's not straightforward to pass the 'Shift' or 'Alt' key status to the painting code where it would be needed)