Introduce PathObjectHierarchy timestamp
Introduce `PathObjectHierarchy.getLastEventTimestamp()`... well, admittedly this was committed in error recently, and this code makes it usable.

The rationale is that sometimes we need to know if the hierarchy is changed, but it is too expensive to add/remove listeners to handle all events.

For example, an expensive lazy measurement that is calculated on demand and cached internally only needs to know whether to recompute the value *at the time when it is requested*. So it can check the object hierarchy timestamp to know whether to do the expensive calculation again, or just to return the last computed value.

Note that, as a side effect of these changes, all detection count measurements are updated automatically when interactively moving an annotation around an image. Previously, these only updated whenever the object was 'committed' to the hierarchy.