Update scatterplot implementation
Reduce the number of objects that need to be created, and use a centralised mouse listener instead of creating a new listener for every node.
Use Circle objects for points instead of styling a StackPane; this makes it possible to bind properties easily, and reduce the number of listeners that need to be added manually.