Remove scatterplot items from start of list
Counterintuitively, this is much faster - especially when removing from large lists (e.g. reducing 100k points to 10k points).
It seems to be because JavaFX is internally removing objects from a long list by reference - and if the item is at the end, then we end up having to search the entire list many times.