Improve ContourTracing efficiency
Partially addresses https://github.com/qupath/qupath/issues/1780 by replacing `Point2` with `IntPoint2`, which stores its values in a packed `long`.

This doesn't make much difference to performance for the ~5 million vertices tracing, but it *did* make it possible to trace much larger regions - making `Polygonizer` the bottleneck.

For example, I could trace two regions - with 39,230,067 and 21,436,882 respectively, in 532 seconds.

Rendering was then slow again, and I couldn't *save* them because of memory issues... but I think it indicates that this PR improves the efficiency of some parts of the tracing.