Add ROI.intersects() method, cache more shapes
The main motivation was that `RegionFilter` was very slow to test when a complex ROI intersected with bounding box. The reason turned out to me that requesting the `Shape` required everything to be copied, and that could be expensive.
This changes makes it possible to do the calculations more quickly, because they don't need to make defensive copies.