Specify number of decimal places for export
Allow the number of decimal places to be specified with `MeasurementExporter`.

Use `MeasurementExporter.DECIMAL_PLACES_DEFAULT` for the default (adaptive) value.

Use a positive number to specify that all non-integer values are export with a fixed number of decimal places.
Use a negative number to specify a maximum number of decimal places.

If the separator is not explicitly requested, any file extension (.csv or .tsv) is now checked to help select the default.