Add icons to single measurement table
The purpose is to:
* Make classifications stand out more (with color)
* Make it possible to distinguish measurements from the measurement list, from those calculated dynamically
* Provide a visual queue to distinguish between annotations, detections & TMA cores

The last of these is done by using the same icons used for show/hide in the toolbar. The hope is that this helps reinforce the association between object type and icon.

Note that the implementation is a bit messy because we don't currently have a way to query the type of a measurement from `ObservableMeasurementTableData`. If/when this is refactored to reduce any JavaFX dependency, we should try to improve this; currently, it's not entirely robust, because someone *could* add measurements or metadata values with duplicate names.