Enable running jpackage with extensions (#1800)
Adds an optional parameter for the jpackage build that takes a
comma-separated list of extensions that should be included in
the build. These must be under the qupath org or the build will fail.

Then, it clones qupath to ./qupath and each extension to
./qupath-extension-whatever, and adds each extension to
[includeBuild] and [dependencies] of include-extra.
Then it runs the gradle build as usual.

For example, you could set the input to
"qupath-extension-instanseg,qupath-extension-djl"