Don't build histogram from large image stacks
QuPath builds a histogram when opening an image, and this is used for Brightness/Contrast settings. This is usually based on a low resolution image and works fine. But it *doesn't* work fine for large z-stacks or time series, especially if these are at a single resolution.

It also isn't fine when opening remote images, e.g. stored through OMERO.

This PR uses a *very* basic approach to select only the middle slice of a z-stack in the event that reading the entire thing would require more than 50 MB to open. This makes it fast to open some images that otherwise throw out-of-memory errors.

For discussion...