Use cached tiles with ImageDisplay histograms
Use whatever low-resolution cached tiles we can find when calculating histograms within `ImageDisplay`.
This is used for auto-setting or resetting the brightness/contrast.

The rationale is:
* We want to construct an `ImageDisplay` quickly, so we don't want to have to visit all the z-slices and time points for a large image just to build a histogram.
* We want to make the best estimate we can of the min/max and overall histogram within the Brightness/Contrast dialog, without requiring that we read an infeasibly large number of pixels.

For this to work, we assume that we're using an `AbstractTileableImageServer`. There's a good chance this performs less well with ImageJ z-stacks, so that behavior will need to be checked.