Reuse tooltips in measurement tables
Tooltips can be reused across nodes.

Applying this to cells of measurement tables can have a significant performance impact, especially if there are large numbers of columns. For a multiplexed image, this took about 1 second off the time required to show the table (although it is still too slow... around 4-5 seconds)