Add min/max local normalization
This applies dilation and erosion to the image to get local minimum and maximum values, then subtracts the minimum and divides by the range.

The minimum and maximum images can also optionally be Gaussian smoothed to avoid sharp discontinuities.

Like other local normalization strategies, this approach can boost the noise in an image if there is no meaningful signal.