Don't show RGB classified tiles if prediction gone
Fixes unintuitive behavior whereby pixel classifier measurements wouldn't appear, despite predictions *seeming* to be made for an image region.

This is because the overlay could merrily show RGB renderings of the prediction whenever the predictions themselves had been ejected from the tile cache... and consequently the predictions were never requested again, since the RGB tile *was* present.

This probably isn't a new bug, and could be replicated by using the memory monitor to clear the tile cache when a pixel classification overlay was visible.

The fix here is to use the prediction image as a key to a weakhashmap used to store RGB tiles (when necessary). The downside is that, when classifying a large image, tiles can start to disappear... although it's probably better than an out of memory error.