Fix 'stuck' points on scatterplot
Workaround to fix https://github.com/qupath/qupath/issues/1877

This sets points to not be visible if their values aren't finite.
It might be preferable to identify points with non-finite values *before* subsampling, so that we don't 'waste' points with values that won't be seen.

But I think this approach is sufficient for now. In the future it would be even better to have charts that can scale to very large numbers of points, and therefore don't require subsampling at all.