Recover if opencv native libs unavailable
The practical benefit is if we wish to make a stripped-down distribution that operates as a viewer but excludes OpenCV and other non-essential features.

We can't remove `qupath-core-processing`, which pulls in OpenCV, but these changes mean that the unavailability of the native libraries doesn't stop QuPath from launching.

This can be tested using
```
./gradlew run -PjavacppPlatform=-
```