Tentative move to Java 25
This 'works', but is likely to require 1 big change and 2 small ones.

The big change is that removing '--bind-services' is probably stripping back the JRE too much, and we'll need to add in more modules or switch to a JDK other than Temurin so that we --bind-services can be reinstated.

The small changes:
- Use Kotlin 2.3 when available
- Use 2.0.0 for org.beryx.runtime when available

The license plugin should also be updated whenever possible, since it currently means we can only use jpackage if the build isn't parallelized.