Unsure pending tile requests are cleared
This was causing a problem whenever pixel classification overlays were stopped in the midst of life prediction: tiles might not be read, and then the exception was retained and returned on every later attempt to access the same tile.

This moves `pendingTiles.remove` to a `finally` block to ensure it is run.

It's possible this causes some issues if a tile really *is* invalid though, so we may need to retain exceptions in some cases - rather than repeatedly making the same (doomed) requests.

Still, this change is necessary to avoid failures in perfectly normal circumstances.